🚚 rename: 요청 함수 폴더 이동
