+++ src/pages/login/components/LoginForm.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import Button from '@/components/common/Button';
-import { useLoginMutation } from './service/postLogin';
+import { useLoginMutation } from '../service/postLogin';
 import router from 'next/router';
 import FormField from '@/pages/signup/components/FormField';
 import Alert from '@/components/dialog/Alert';
